// studio/src/lib/types.ts
export type DialogAction = {
  action: string;
  node?: string;
  opts?: string[];
};

export type DialogChoice = {
  id: string;
  value: string;
  next: DialogAction[];
};

export type DialogNode = {
  id: string;
  isSpeak?: boolean;
  isPrompt?: boolean;
  bubbles: string[];
  before: DialogAction[];
  next: DialogAction[];
  choices: DialogChoice[];
};

export type DialogScript = {
  id: string;
  first: string;
  nodes: DialogNode[];
};